Geolocation works, but still WIP
I updated code work to correctly with geo_location. I've now tested this code
to create/delete generic, alias and geo_location records, and it works as
expected.

The main issue is that the code does a `rescue` at line 156 of
`providers/record.rb`, because it's not properly test-and-repair, so it
rescues when you DELETE a record that doesn't exist, but also when you try to
update records that aren't in your zone.

I'm committing as a WIP with this caveat, but it's completely usable as is.
I'll come back to see what's going on with the `current_resource_record_set`
comparison.

ALSO, deleted the ChefSpec `test_spec.rb` since it wasn't of any use.